backend/pyproject.toml
[project]
name = "stark-vix-backend"
version = "0.1.0"
description = "STARK-VIX multi-agent research assistant — FastAPI backend"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.6",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "sse-starlette>=2.1.3",
    "httpx>=0.27.0",
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.1",
    "pymupdf>=1.24.0",
    "python-multipart>=0.0.9",
]

[dependency-groups]
dev = [
    "ruff>=0.7.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]

[project.optional-dependencies]
# STARK-VIX agent stack. Instalar con: uv sync --extra agents
# LangGraph orquesta; cada provider es opcional en tiempo de uso — la factory
# resuelve dinámicamente (get_chat_model("ollama", "gemma2:27b") etc.).
agents = [
    "langgraph>=0.2.50",
    "langchain>=0.3.7",
    "langchain-core>=0.3.15",
    "langchain-ollama>=0.2.0",
    "langchain-anthropic>=0.2.4",
    "langchain-openai>=0.2.8",
    "langchain-google-genai>=2.0.4",
    "umap-learn>=0.5.6",
    "scikit-learn>=1.5.0",
    "google-generativeai>=0.8.3",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
